mongodb changed to postgresql in payment-srv
