+++ payment/prisma/schema.prisma
@@ -1,0 +1,18 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model Order {
+  id             Int      @id @default(autoincrement())
+  userId         String
+  courseId       String
+  price          Float
+  payment_status String   @default("incomplete")
+  createdAt      DateTime @default(now())
+  updatedAt      DateTime @updatedAt
+}
