--- payment/src/@types/modelTypes/order.ts
@@ -1,6 +1,0 @@
-export interface IOrder {
-  userId?: string;
-  courseId: string;
-  price?: number;
-  payment_status?: string;
-}
