+++ payment/src/config/prismaClient.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from '@prisma/client'
+
+const prisma = new PrismaClient()
+
+export default prisma
